// server/src/types/Catalog.ts
export interface PersonType {
  codigo: number;
  descripcion: string;
}
// ...
export interface Person {
  codigo: number;
  nombre: string;
  apellido: string;
  telefono: string;
  celular: string;
  direccion: string;
  activo: string;
  correo: string;
  ruc: string;
  documento: string;
  digitoVerificador: string;
  observacion?: string | null;
  estadoCivil: string;
  fechaNacimiento: string;
  tipoDocumentoCodigo: number;
  tipoDocumentoDescripcion: string;
  fechaGrabacion: string | null;
}

export interface PersonDetail {
  persona: Person;
  assignedTypes: PersonType[];
  availableTypes: PersonType[];
}

export interface PersonInput {
  nombre: string;
  apellido: string;
  telefono: string;
  celular: string;
  direccion: string;
  activo: string;
  correo: string;
  ruc: string;
  documento: string;
  digitoVerificador: string;
  observacion: string | null;
  estadoCivil: string;
  fechaNacimiento: string;
  tipoDocumentoCodigo: number;
}